requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
docopt==0.6.2
frozenlist==1.4.0
geographiclib==2.0
geopy==2.4.0
grpcio==1.59.0
grpcio-tools==1.59.0
idna==3.4
importlib-metadata==6.8.0
ipinfo==4.4.3
multidict==6.0.4
pipreqs==0.4.13
platformdirs==3.11.0
protobuf==4.24.4
psutil==5.9.6
py-cpuinfo==9.0.0
requests==2.31.0
tomli==2.0.1
urllib3==2.0.6
yapf==0.40.2
yarg==0.1.9
yarl==1.9.2
zipp==3.17.0
